pyproject.toml
[project]
name = "rms-app-ds"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "sqlmodel>=0.0.27",
    "psycopg[binary,pool]>=3.2.11",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "playwright>=1.58.0",
    "ty>=0.0.12",
]

[tool.apx.metadata]
app-name = "rms-app-ds"
app-slug = "rms_app_ds"
app-entrypoint = "rms_app_ds.backend.app:app"
api-prefix = "/api"
metadata-path = "src/rms_app_ds/_metadata.py"

[tool.apx.ui]
root = "src/rms_app_ds/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.version]
path = "src/rms_app_ds/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/rms_app_ds/__dist__", "src/rms_app_ds/_metadata.py"]
exclude = ["src/rms_app_ds/ui"]
